Fix Docker kill bug
We need to send a signal that isn't interpreted by the docker run process.
Otherwise, a race condition can result where the signal is sometimes caught
by the trap, and sometimes caught by the docker process.

SIGCONT serves this purpose, since it is ignored by Docker.
Note that SIGUSR[12] is not ignored, much to my surprise.